commit on master

+++ loops/test.py
@@ -173,8 +173,8 @@ def test(cfg):
                 #obj_pred_choice = obj_pred_choice.view(-1,1)
                 #obj_target = obj_target.view(-1,1)
                 #np.save(data['dir']+'/streamlines_lstm_GIN',streamlines)
-                mae = torch.mean(abs(obj_target.data.cpu() - obj_pred_choice.data.cpu())).item()
-                mse = torch.mean((obj_target.data.cpu() - obj_pred_choice.data.cpu())**2).item()
+                mae = torch.mean(abs(obj_target.data.float().cpu() - obj_pred_choice.data.cpu())).item()
+                mse = torch.mean((obj_target.data.float().cpu() - obj_pred_choice.data.cpu())**2).item()
                 #correct = obj_pred_choice.eq(obj_target.data.int()).cpu().sum()
                 #acc = correct.item()/float(obj_target.size(0))
